Add a script to clean the FMC's configuration
It appears that the lab image contains the `DNE Security Access Control Policy` that the example script is supposed to create.  This causes the script to error-out (since a policy with the same name already exists).

We should probably fix the lab image; however in the mean time, the `fmc_clean_access_policy.py` script will check for the policy and delete it, if found.